% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_auc.R
\name{max_auc}
\alias{max_auc}
\title{Estimate the Maximum AUC}
\usage{
max_auc(k, pve, n.bins = 100)
}
\arguments{
\item{k}{(numeric; range 0-1) Population prevalence of the disease of interest}

\item{pve}{(numeric; range 0-1) Proportion of variability explained by the PRS}

\item{n.bins}{(integer) Number of bins for estimating the AUC (default = 100)}
}
\value{
A numeric value for the maximum AUC
}
\description{
Estimate the maximum AUC of a polygenic risk score based on the population prevalence of a disease \code{k} and the proportion of variability explained by the PRS \code{pve}.
}
\examples{
max_auc(k = 0.13, pve = 0.26)
}
