% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_se.R
\name{opt_se}
\alias{opt_se}
\title{Optimal Sensitivity}
\usage{
opt_se(
  k,
  pve,
  sp,
  n.bins = 1000,
  thresh.vector = seq(from = 1, to = n.bins, by = 10),
  direction = "max"
)
}
\arguments{
\item{k}{(numeric; range 0-1) Population prevalence of the disease of interest}

\item{pve}{(numeric; range 0-1) Proportion of variability explained by the PRS}

\item{sp}{(numeric; range 0-1) Specificity}

\item{n.bins}{(integer) Number of bins for estimating the AUC (default = 100)}

\item{thresh.vector}{(numeric) Vector of thresholds}

\item{direction}{(character) \code{"max"} or \code{"min"}}
}
\value{
A numeric valuve for the optimal sensitivity
}
\description{
These functions estimate the optimal (maximal or minimal) sensitivity of a PRS.
}
\examples{
opt_se(k = 0.13, pve = 0.26, sp = 0.99)
}
